Refactor local import task into domain services
